Restructure test packages
