+++ jaxrs/src/test/java/se/fortnox/reactivewizard/jaxrs/ExceptionHandlerTest.java
@@ -1,7 +1,7 @@
 package se.fortnox.reactivewizard.jaxrs;
 
 import se.fortnox.reactivewizard.ExceptionHandler;
-import se.fortnox.reactivewizard.MockHttpServerResponse;
+import se.fortnox.reactivewizard.mocks.MockHttpServerResponse;
 import io.netty.buffer.ByteBuf;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.reactivex.netty.protocol.http.server.HttpServerRequest;
