+++ jaxrs/src/test/java/se/fortnox/reactivewizard/jaxrs/SynchronousResourcesTest.java
@@ -7,15 +7,16 @@ import io.reactivex.netty.protocol.http.server.RequestHandler;
 import org.junit.Test;
 import rx.Scheduler;
 import rx.schedulers.Schedulers;
+import se.fortnox.reactivewizard.utils.JaxRsTestUtil;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import java.util.concurrent.Executors;
 
-import static se.fortnox.reactivewizard.jaxrs.JaxRsTestUtil.testServer;
+import static se.fortnox.reactivewizard.utils.JaxRsTestUtil.testServer;
 import static org.fest.assertions.Assertions.assertThat;
 
-public class SynchronousResources {
+public class SynchronousResourcesTest {
 
     @Path("threadname")
     public class TestRes {
