+++ jaxrs/src/test/java/se/fortnox/reactivewizard/jaxrs/WrappedParamTest.java
@@ -1,7 +1,7 @@
 package se.fortnox.reactivewizard.jaxrs;
 
 import se.fortnox.reactivewizard.ExceptionHandler;
-import se.fortnox.reactivewizard.MockHttpServerResponse;
+import se.fortnox.reactivewizard.mocks.MockHttpServerResponse;
 import io.netty.handler.codec.http.HttpMethod;
 import io.reactivex.netty.protocol.http.server.MockHttpServerRequest;
 import org.junit.Test;
